+++ ChangeLog.md
@@ -1,3 +1,7 @@
+### 2026-07-20 Version 1.0.3
+* feat: add DoSSEActionAsync and ReadAsSSEAsync (netstandard2.1+)
+* feat: add netstandard2.1 target framework
+
 ### 2026-07-16 Version 1.0.2
 * fix: match retryCondition.Exception by exception type name, not Message